+++ src/extension.ts
@@ -1,24 +1,13 @@
 import * as vscode from 'vscode';
-import { welcomeMessage } from './extensions/welcome-command';
+import { welcomeMessage } from './extensions/welcomeMessage';
+import { registerStatusBar } from './extensions/statusBar';
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
-
-	// Use the console to output diagnostic information (console.log) and errors (console.error)
-	// This line of code will only be executed once when your extension is activated
-	console.log('Congratulations, your extension "pray-time" is now active!');
-
-	// The command has been defined in the package.json file
-	// Now provide the implementation of the command with registerCommand
-	// The commandId parameter must match the command field in package.json
-	const disposable = vscode.commands.registerCommand('pray-time.welcome-message', () => {
-		// The code you place here will be executed every time your command is executed
-		// Display a message box to the user
-		vscode.window.showInformationMessage('Welcome to Pray Time extension!');
-	});
-
+	registerStatusBar(context);
 	welcomeMessage(context);
+
 
 	// context.subscriptions.push(disposable);
 }