+++ src/extensions/welcomeMessage.ts
@@ -1,9 +1,10 @@
 import { ExtensionContext, window, workspace } from "vscode";
+import useDate from "../hooks/useDate";
 
 export const welcomeMessage = (context: ExtensionContext) => {
     const enableMessage = workspace.getConfiguration().get<boolean>('pray-time.welcome-message.enabled');
     if(enableMessage) {
-        let city = workspace.getConfiguration().get<string>('pray-time.welcome-message.city');
-        window.showInformationMessage(`Hello ${city || 'You'} Welcome to Pray Time Extension!`);
+        let message = useDate();
+        window.showInformationMessage(`Welcome to Pray Time Extension! Today is ${message}`);
     }
 };