+++ src/hooks/useLocalization.ts
@@ -1,0 +1,12 @@
+const useLocalization = (city: string) : string => {
+    if(city === 'Jakarta') {
+        return 'ID';
+    } else if(city === 'London') {
+        return 'UK';
+    }
+    else {
+        return 'ID';
+    }
+};
+
+export default useLocalization;