+++ CHANGELOG.md
@@ -8,3 +8,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Initial release
 - Add feature: Welcome Message & Configuration
+- Add feature: Countdown Pray Time (Status Bar)